run VM using qemu and provide it with RPM repo
This patch introduces few functions to run a VM using qemu and serve RPM
packages over HTTP using the Python's built-in HTTP server. The
configuration format is currently very simple yaml file like this:
```
rpms-directory: "rpms"
image: "Fedora-Cloud-Base-32-1.6.x86_64.qcow2"
```
Both paths are relative to the location of the configuration file.

The RPM repo is configured (in the guest) using cloud-init. A pair of
ephemeral SSH keys is also created in the cache directory.