expect the input RPM to contain the setup logic
take the name of RPMs for test and target VM from the configuration
file. Expect the test RPM to contain whatever setup logic is required
for both the test and target VMs.

Also switch to JSON as it is simpler and comes with better tooling and
in general it does not really matter. Example configuration:
```
{
	"rpms-directory": "rpms",
	"image": "Fedora-Cloud-Base-32-1.6.x86_64.qcow2",
	"test-rpm": "osbuild-composer-tests",
	"rpmci-setup": "/usr/libexec/osbuild-composer/rpmci-setup",
	"tests-directory": "/usr/libexec/tests/osbuild-composer/",
	"target-rpm": "osbuild-composer"
}
```