rpmci/configuration: replace hand written parser with jsonschema
The previous implementation was growing too fast, usage of jsonschema
will make it more readable and it can also be easily used as a source of
example configurations which was not the case previously.

This patch also changes the configuration format because we had few
duplicates in there, namely:
 * command invocation was available twice
 * credentials could be specified at two places

The disadvantage of this patch is that rpmci fails with jsonschema
internal exception in case something goes wrong, but this can easily be
fixed by handling it.