rpmci: introduce a way to pass credentials to the guest VM
I call it guest features, because naming ... The idea is to pass all the
information that the test case might need, for example: aws credentials,
nested virtualization, access to RHEL content, etc.

So far it simply takes the input JSON object and passes it to the invoke
command through stdin.